% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_dfs.R, R/base_aov_dfs.R,
%   R/containment_dfs.R, R/inner_outer_dfs.R, R/nlme_dfs.R,
%   R/satterthwaite_dfs.R
\name{residual_aov}
\alias{residual_aov}
\alias{asymptotic_aov}
\alias{data_aov_dfs}
\alias{containment_aov}
\alias{inner_outer_aov}
\alias{nlme_aov}
\alias{satterthwaite_aov}
\title{Degrees of Freedom Calculations}
\usage{
residual_aov(model = model, type = type)

asymptotic_aov(model = model, type = type)

data_aov_dfs(formula, data)

containment_aov(model = model, type = type)

inner_outer_aov(model = model, type = type)

nlme_aov(model = model, type = type)

satterthwaite_aov(model = model, type = type)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}

\item{type}{type of test, \code{II, III, 2, 3}. Roman numerals are equivalent
to the corresponding Arabic numerals.}
}
\value{
a \code{data.frame}
}
\description{
Methods for calculating degrees of freedom:

\code{residual_aov}: DFs are calculated from \code{df.residual}
\code{asymptotic_aov}: DFs are set to \code{Inf} for all model terms
\code{nlme_aov}: DFs are calculated based on the algorithm given in Mixed-Effects Models in S and S-Plus and employed in the \code{nlme} package
\code{containment_aov}: DFs are calculated using the containment algorithm
\code{satterthwaite_aov}: DFs are calculated using the Satterthwaite approximation, as implemented in \code{glmmTMB::dof_satt}
}
